modified for sticky topbar and added application placeholder name
