+++ client/index.html
@@ -3,13 +3,14 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>COM</title>
+    <title>WarfighterFit</title>
     <script>
       (function () {
         var theme = localStorage.getItem("setting_theme") || "nightshade";
         var compact = localStorage.getItem("setting_compact_cards") === "true";
         document.documentElement.setAttribute("data-theme", theme);
-        if (compact) document.documentElement.setAttribute("data-compact", "true");
+        if (compact)
+          document.documentElement.setAttribute("data-compact", "true");
       })();
     </script>
   </head>
